refactoring release step of functions deploy to use typescript
