+++ src/deploy/functions/index.ts
@@ -1,4 +1,3 @@
-import * as release from "./release";
 export { prepare } from "./prepare";
 export { deploy } from "./deploy";
-export { release };
+export { release } from "./release";
